ptune: Release v1.5.0
Still not v2 yet, but you're gonna want this one. My dev environment made it hard to realize that the Pixels start SurfaceFlinger earlier in the boot chain for a smoother transition to the boot animation earlier on, earlier in relation to when Magisk runs my post-fs-data service. The temporary fix for now is to restart SurfaceFlinger, which gives about a 4 second increase in boot time running on my Pixel 6 Pro. The intended future fix is to set the SF props earlier in the init process while still overriding the values set by the ROM, but with my new job I need some more time before I can sort this out.

service:
- Restart SurfaceFlinger after changing its props
- Change foreground_window group (hOS exclusive) from all cores to little and mid cores (as for gs101 layout)
- Change foreground group from little and mid cores to all cores (as for gs101 layout)

I make mention of the gs101 layout, as with the continued delays of the v2 release we are still operating on the basis of this layout, as it's the only device I have available to test. Once v2 is ready, the core layouts will be appropriately defined for my "PowerPulse" service to calculate the correct cores to assign to each group according to my rule definitions, with user customizability via an app. Until then, I apologize 🫠

This release also drops support for Android 14, as I don't believe that it works anymore after the fixes required for the November security patch on Android 15. Again, nearly all versions of Android should be supported once v2 is ready to roll out, so I don't consider this to be a setback at the moment d: